+++ tests/test_vesq.py
@@ -4,7 +4,7 @@ from telliot_core.apps.core import TelliotCore
 from web3.datastructures import AttributeDict
 
 from telliot_feeds.feeds.vesq import vsq_usd_median_feed
-from telliot_feeds.reporters.tellorflex import TellorFlexReporter
+from telliot_feeds.reporters.tellor_flex import TellorFlexReporter
 
 
 @pytest.mark.asyncio
